// src/api/generated/model/updateProductCommand.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication and Versioning
 * OpenAPI spec version: v1
 */

export interface UpdateProductCommand {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  price?: number;
  categoryId?: string;
}